Updated version to proper semantic versioning
As this is a early stage alpha/beta of this addon, and functions are regularly added, there are some things not done yet. That, specifically, would be using saved variables, through slashcmd /st, we need to look up potential conflicts with /st though. If it's something popular, a guard for it needs adding.